added a domain folder. created types file. moved util to domain folder
