+++ src/RPGTools/TreePreview/HistoryItem.tsx
@@ -1,6 +1,6 @@
 
 import "./history-item.css";
-import { DialogueHistory } from "./TreePreviewPage";
+import { DialogueHistory } from "./domain/types";
 
 type Props = {
     history: DialogueHistory;
